+++ effectful-th/CHANGELOG.md
@@ -3,6 +3,8 @@
 * Don't generate signatures with an out-of-scope name when a constructor
   mentions the monad variable in its context (e.g. `Op :: Monad m => Int -> E m
   ()`) or in arguments of the effect type (e.g. `Op :: Int -> E (m Int) m ()`).
+* Correctly transfer fixity annotations of constructors to the generated
+  functions.
 
 # effectful-th-1.0.0.3 (2024-10-08)
 * Make `makeEffect` reuse Haddock descriptions of effect operations for